Update Onboarding.ps1
Inserted the smtp port 587 for the smpt server to be utilized.
Inserted username variable and password variable so these could be stored in the $pscred variable so the input would not need user input on the email sending. this should allow a smooth transition for the email to send with all inputs without a box popping up for cred input.